Move TestSummarizeUserCounts
